pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-genesis-env"
version = "0.1.0"
description = "GENESIS — The Autonomous Startup Gauntlet. Train LLMs to co-found and operate a startup from Day 0 to Series A."
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "fastmcp>=0.1.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "genesis_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["genesis_env", "genesis_env.server"]
package-dir = { "genesis_env" = ".", "genesis_env.server" = "server" }
